+++ tests/integration/test_run_component/test_script.py
@@ -1,5 +1,13 @@
 import pytest
 import sys
+from io import StringIO
+from contextlib import redirect_stdout
+
+### VIASH START # noqa: E266
+meta = {
+    "memory_gb": 6,
+}
+### VIASH END # noqa: E266
 
 
 def test_run_component(run_component):
@@ -14,4 +22,17 @@ def test_run_component(run_component):
 
 
 if __name__ == "__main__":
-    sys.exit(pytest.main([__file__], plugins=["viashpy"]))
+    temp_stdout = StringIO()
+    print(f"meta: {meta}")
+    assert (
+        meta["memory_gb"] is not None
+    ), "This test should be executed with some --memory set."
+    with redirect_stdout(temp_stdout):
+        result = pytest.main([__file__], plugins=["viashpy"])
+    stdout_str = temp_stdout.getvalue()
+    print(stdout_str)
+    assert (
+        "Different values were defined in the 'meta' dictionairy that limit memory, choosing the one with the smallest unit"
+        not in stdout_str
+    )
+    sys.exit(result)